Use plain UUIDs for unique names
The types which implement the Namer interface were prefixing their name
with their resource type. This is problematic, because it does not
comply with the allowable character set for kubernetes labels.

To allow their names to be easily used in labels, this commit uses the
UUID alone as the name. The full resource name is now available through
the ResourceName method.

In addition, the Kubernetes Namer interface has been updated to become
ResourceNamer, encapsulating the new ResourceName method.