Add preliminary thread-safe availability queue
This change adds a custom Queue type. It uses a ReservationLimiter to
only dequeue sessions that can be scheduled with the machines that are
currently available.

Note this commit does not remove the kubernetes queue in the controller.
This only implements a new one. The replacement will happen in a later
commit.