Add pool discovery function
This commit moves the definition of the Pool type from the
reservation.go file into pool.go. Then, it implements a FindPools
function that can create Pool instances with correct capacities by
querying the kubernetes API for nodes.

This commit does not use the discovery as part of the startup process.
This will come in a later commit.