Add driver logs to events and rename event message
This change adds a list of bytes containing the driver logs to the event
type and the TestSessionResult message. A slice of bytes is chosen,
because the format of the logs is not intended to be understood by the
services. This is also the way kubernetes returns them from the pod. It
is up to the reader to parse the logs. Since []byte prevents direct
comparison of event types, this change switches tests to use
reflect.DeepEqual.

In addition, this change renames the `Message` field on the Event struct
to `Description` and adds it to the Event message.