Stop clean-up error returns in Executor
The Execute method of the Executor is mainly responsible for calling the
appropriate methds to provision, monitor and clean-up a session.
Previously, it returned an error when something went wrong in any of
those stages.

Unlike the other stages, the clean-up stage occurs when the previous
were successful or errored. An error during clean-up was not indicative
of an error during the test. For this reason, this change switches the
clean-up stage to log error messages. It no longer returns a clean-up
error.