Add tests for Executor
The Executor orchestrates a single test session by managing calls to
provision and clean-up test components. Meanwhile, it monitors their
status.

Users should use an Executor through the Execute method, which invokes a
series of unexposed methods. This commit adds tests for two of these
methods: provision and monitor. These methods create the kubernetes pods
for the components and monitor the status of them during test execution.

Unfortunately, the clean-up method is harder to test. Kubernetes exposes
a fake to test, but this fake does not support the functionality used by
Executor's clean method. See kubernetes/client-go#609. For this reason,
the test is commented out in this version.

Finally, the Execute method itself is not yet tested. Since it relies on
these unexposed functions which are easier to test, it's not a high
priority.