Refactor controller and executor, add/improve tests
This commit restructures the Controller type to rely more heavily on
dependency injection and smaller methods. It adds tests that make use
this simplified structure.

In addition, executor is renamed to kubeExecutor. A new Executor
interface is introduced. The controller now relies on this interface,
decoupling it from the kubeExecutor. This allows the controller to mock
interactions with exectors.

Finally, test logic, organization and mocking is improved throughout the
orch package.