Add nil clientset check in Controller constructor
The kubernetes clientset should not be nil. While this case seems
unlikely, it could result in a panic during the construction of a
Controller. This commit adds an error to the returned values in the
NewController function. If the clientset is nil, an error is returned.

In addition, the commit introduces tests to ensure this behavior is
retained across future changes and modifies the executables to handle or
ignore the construction error.