Fix blocking issue with multiple sessions
When multiple sessions were run serially, the second would block
forever. The Watcher type sends Kubernetes events for a specific
session over a channel to the executor. During clean-up the Watcher
continued to post events, but the executor never read these. Since the
Watcher used an unbuffered channel, it blocked until it was read. This
blocking caused a lock to never be released. This lock was needed to
create a new channel for a new session.

This change switches the Watcher to use a buffered channel which does
not block until it is full. The Watcher is also modified to drop events
when the channel's capacity is exceeded. This prevents it from blocking
during manual tests and allows additional sessions to be scheduled.

In addition, this change adds a missing call to (*queue).Done in the
controller. This call is necessary to instruct the reservation system
that machines are available for scheduling.