Use a context parameter in (*Controller).Stop
The Stop method of the orch.Controller type accepted a time.Duration. It
would wait for a graceful shutdown for this amount of time. Then, it
would return an error if the graceful shutdown did not occur.

This commit replaces the time.Duration with a context.Context,
permitting more flexibility.